Truncating session names for role assumption
Session names have a poorly-documented 64 char limit, which can cause
role assumption to fail when long role names are used.

I've also moved the hash to the front of the session name because it
simplifies chopping off the end of the role name when necessary. That's
based on a very biased assumption that role name prefixes are more
informative than suffixes.