We run a number of cronjobs in our environment.
These pods finish running in a Complete or Failed state and linger for sometime. Unfortunately kubernetes leaves the PodIP intact that was used for these pods.

If docker reuses these ips then we have a conflict and this leads to ips being flagged against the wrong namespace.

This change restricts the namespace tracking to pods in a Running phase.

This might not be a complete solution to this issue.