Limit relevant pods to those on the same node as kube2iam
In clusters with 1000's of pods the memory used by each kube2iam
instance explodes because each kube2iam instance is storing information
about all pods on all nodes in memory.

This adds a flag `--node` for specifying the node name of kube2iam. If
set, it will limit the pods it's watching to those only on the same
node as the kube2iam instance.