:bug: Remove remote address header when any IMDSv2 tokens are used
Even after #270 we have experienced 403s from the metadata service while
using the AWS VPC CNI with any pods not using host networking. This
indicated that while the PUT requests for tokens were now working
correctly, the subsequent GET requests still had the `X-Forwarded-For`
header which AWS rejects with a 403.