Fix issue with CircleCI build.
The issue was that the Clientset struct, which contains the CoreV1() method, was embedded directly in the Client struct. While this works in some Go versions, it seems the CircleCI environment's Go version requires explicit access to the embedded struct. I've updated the code to use k8s.Clientset.CoreV1() to explicitly access the CoreV1() method through the embedded Clientset. This should resolve the build and test failures in CircleCI.